+++ packages/jest-snapshot/src/__tests__/inline_snapshots.test.ts
@@ -6,7 +6,7 @@
  */
 
 jest.mock('graceful-fs', () => ({
-  ...jest.genMockFromModule<typeof import('fs')>('fs'),
+  ...jest.createMockFromModule<typeof import('fs')>('fs'),
   existsSync: jest.fn().mockReturnValue(true),
   readdirSync: jest.fn().mockReturnValue([]),
   statSync: jest.fn(filePath => ({